Fully functional 100% PB-compatible map upload!

+++ src/map/location.hpp
@@ -12,7 +12,7 @@
 
 const uint8_t LOCATION_WIDTH = 32;
 const uint8_t LOCATION_HEIGHT = 18;
-const uint8_t LOCATION_BLOCKS_NUMBER = LOCATION_WIDTH*LOCATION_HEIGHT;
+const uint32_t LOCATION_BLOCKS_NUMBER = LOCATION_WIDTH*LOCATION_HEIGHT;
 
 class Location : public sf::Drawable, public sf::Transformable {
 public:
